Add filter by date
