+++ types/logs.ts
@@ -1,0 +1,7 @@
+export type Log = {
+  id: number;
+  date: Date;
+  itemName: string;
+  servings: number;
+  thumbnail: string;
+};
